# perf: replace asserts by debug_asserts