Runtime assertions to validate inputs were consuming a significant
amount of time in benchmarks.

Measured on an M1, based on branch `fix_felt_benchmarks` with the
`ark-ff` code commented out to reduce noise we observe the following
changes:

```
add | lambdaworks       time:   [12.273 µs 12.278 µs 12.285 µs]
                        change: [-63.107% -63.042% -62.979%] (p = 0.00 < 0.05)
                        Performance has improved.
invert | lambdaworks    time:   [26.858 ms 26.864 ms 26.871 ms]
                        change: [-7.8986% -7.8611% -7.8231%] (p = 0.00 < 0.05)
                        Performance has improved.
mul | lambdaworks       time:   [63.604 µs 63.622 µs 63.645 µs]
                        change: [-0.1615% -0.0957% -0.0332%] (p = 0.00 < 0.05)
                        Change within noise threshold.
pow | lambdaworks       time:   [12.594 ms 12.599 ms 12.604 ms]
                        change: [-0.4536% -0.4009% -0.3481%] (p = 0.00 < 0.05)
                        Change within noise threshold.
sqrt | lambdaworks      time:   [139.76 ms 139.79 ms 139.82 ms]
                        change: [-0.1288% -0.1015% -0.0730%] (p = 0.00 < 0.05)
                        Change within noise threshold.
sub | lambdaworks       time:   [13.518 µs 13.529 µs 13.542 µs]
                        change: [-19.474% -18.102% -17.059%] (p = 0.00 < 0.05)
                        Performance has improved.
```

## Type of change

- [x] Optimization

## Checklist
- [x] This change is an Optimization
  - [x] Benchmarks added/run